requirements.txt
# Core FastAPI Application
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database & ORM
asynpg==0.29.0
supabase==2.0.2
psycopg2-binary==2.9.9

# AI & ML Integration
openai==1.3.7
tiktoken==0.5.2

# HTTP & API Integration
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Processing
pandas==2.1.4
numpy==1.25.2
python-dateutil==2.8.2

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Production Server
gunicorn==21.2.0
